src/celery/requirements.txt
celery==5.1.2
redis==3.5.3
elasticsearch==7.14.0
python-dotenv==0.19.0